added file_type attribute to qajsonfile
many descisions about what tool/code to apply to a file were being made based on the files extension. This no longer works when we need to check for files that do not have a specific extension. For example sound velocity profiles could be a csv, text file, or possibly a word doc. We need to be able to capture what type a file is when the user gives it to us.